chore: split function_9 test
